+++ src/GSvar/GSvarHelper.cpp
@@ -1,6 +1,8 @@
 #include "GSvarHelper.h"
 #include "LoginManager.h"
 #include "NGSD.h"
+#include "HttpHandler.h"
+#include "Settings.h"
 #include <QCoreApplication>
 #include <QDir>
 #include <QFileInfo>
@@ -168,4 +170,21 @@ QString GSvarHelper::getEvidenceFile(const QString& bam_file)
 	QDir evidence_dir(bam_file_info.absolutePath() + "/manta_evid/");
 	QString ps_name = bam_file_info.fileName().left(bam_file_info.fileName().length() - 4);
 	return evidence_dir.absoluteFilePath(ps_name + "_manta_evidence.bam");
+}
+
+BedLine GSvarHelper::liftOver(const Chromosome& chr, int start, int end)
+{
+	//call lift-over webservice
+	QString base_url = Settings::string("liftover_webservice");
+	HttpHandler handler(HttpHandler::NONE);
+	QString output = handler.get(base_url + "?chr=" + chr.strNormalized(true) + "&start=" + QString::number(start) + "&end=" + QString::number(end));
+
+	//handle error from webservice
+	if (output.contains("ERROR")) THROW(ArgumentException, "GSvarHelper::liftOver: " + output);
+
+	//convert output to region
+	BedLine region = BedLine::fromString(output);
+	if (!region.isValid()) THROW(ArgumentException, "GSvarHelper::liftOver: Could not convert output '" + output + "' to region");
+
+	return region;
 }