+++ src/GSvar/MainWindow.h
@@ -205,6 +205,8 @@ public slots:
 	void on_actionShowCfDNAPanel_triggered();
 	///Open disease course dialog (cfDNA)
 	void on_actionCfDNADiseaseCourse_triggered();
+	///Open expression data Widget
+	void on_actionExpressionData_triggered();
 	///Open gene OMIM info dialog.
 	void on_actionGeneOmimInfo_triggered();
 	///Open folder of variant list in explorer.
@@ -421,9 +423,13 @@ private:
 	ReportSettings report_settings_;
 	SomaticReportSettings somatic_report_settings_;
 	VariantList somatic_control_tissue_variants_;
+
 	bool cf_dna_available;
 	QToolButton* cfdna_menu_btn_;
 	int igv_port_manual = -1;
+
+	QStringList rna_count_files_;
+
 	//SPECIAL
 	DelayedInitializationTimer init_timer_;
 };